refactor(core): extract Pep isa3 core into subdirectory
That way the RISC-V core implementation is sibling of Pep rather than being contained within the same dir.
