+++ core/core/sim/cores/cpu/pep/pep_isa.cpp
@@ -4,7 +4,7 @@
 #include "core/arch/pep/isa/pep10.hpp"
 #include "core/arch/pep/isa/pep9.hpp"
 #include "core/ds/string_compare.hpp"
-#include "core/sim/cores/cpu/pep_isa_instructions.hpp"
+#include "core/sim/cores/cpu/pep/pep_isa_instructions.hpp"
 #include "core/sim/debugger/register_scanner.hpp"
 #include "core/sim/memory/ram/dense.hpp"
 #include "core/sim/system.hpp"
