+++ test/core/sim/core/pep/instr/api.hpp
@@ -1,8 +1,8 @@
 #pragma once
 #include "core/arch/pep/isa/pep10.hpp"
 #include "core/arch/pep/isa/pep9.hpp"
-#include "core/sim/cores/cpu/pep_isa.hpp"
-#include "core/sim/cores/cpu/pep_isa_instructions.hpp"
+#include "core/sim/cores/cpu/pep/pep_isa.hpp"
+#include "core/sim/cores/cpu/pep/pep_isa_instructions.hpp"
 #include "core/sim/memory/ram/dense.hpp"
 #include "core/sim/system.hpp"
